Add functions to fpdf_annot.h to provide access to "Opt" dictionary size and option labels by index.
Part of feature work to make Pdfium more usable by mobile platforms.
Adds 2 methods to fpdf_annot.h:
    FPDFAnnot_GetOptionCount - Returns the number of entries in
    annotation's "Opt" array if present. Else -1.

    FPDFAnnot_GetOptionLabel - Gets the string value for the label of
    the option at a given index in the annotation's "Opt" array.
    Returns the length of the string value or -1 if annotation does not
    have an "Opt" array or index is invalid.

This work is intended to be used by mobile clients that wish to render
comboboxes and listboxes using custom displays when a widget is
selected instead of Pdfium's support for drawing the dropdown/selected
items.

R=thestig@chromium.org

Bug: 124252477
Change-Id: I4b91a43febbc7dd4f7ce46f744424ddbf8af50f8