Handle creation of CGFAS_Decimals from floats with bigger values.
Alas, the current implementation is 96 bits; we'd need 128 to cover
the full range of an IEEE float.

Dividing a float by 1e32 does not produce a 32-bit shift. The current
code "works" because 64 bits are accidentally carried into |plo| with
|phi, pmid| getting 0s (mercifully) from the botched divisions (except
in some large cases) and then redistributed.

Beef up some tests while we're at it.

Change-Id: I01922d368c52008f9e242909820afeaae796da5f