Update patch set 1
Patch Set 1:

This is a slightly larger cl and libopenjpeg2 in pdfium is still patched, so I expect some discussion.
The main motivation for this is to remove the need to manually unbundle libopenjpeg2 downstream, as this is error-prone (patches break with each Chromium release) and Chromium itself does not provide an unbundling mechanism as it does for other libraries.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 6780 <6780@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false