Update patch set 3
Patch Set 3:

Already covered by test CL that landed the feature, as in:

https://cs.chromium.org/chromium/src/third_party/pdfium/testing/resources/javascript/mouse_events_expected.txt?rcl=e6fcdfa5bb6e1fca1bb833cbace54b754ca395a6&l=8

Recall that we added the ability for pdfium_test to read a .evt file and send simple events so that we wouldn't have to write an embedder test for every one of these interactive tests.

Is there some other coverage that this new test adds that I'm missing?

Patch-set: 3
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>