Avoid crashing inside CPDF_DIBBase.
Currently, CPDF_ImageRenderer::Continue() can fail in |Mode::kDefault|,
and then attempt to restart rendering. In valid uses of this code path,
the rendering mode switches to |Mode::kBlend|. In the crashing case, the
rendering mode remains as |Mode::kDefault|. Avoid this crash by checking
if the rendering mode stays the same and refusing to continue.

BUG=pdfium:1219

Change-Id: I232cb4da2a248f923f09de7d4f08b30ae7dde832