Fix another CHECK() in CFGAS_StringFormatter::FormatNum()
This was broken in f234dde95a which introduced a length calculation
which was off by one. Additionally, it looks like the calculation
would be wrong since we have the position inside the new string which
is concatenated with the old one at line xxxx, so the position needs
adjustment based on the length of the old string.

Bug: pdfium:1271
Change-Id: I3de697bfc29f7dc5b133e5bc7c08f574ff9eeebd