Make CXFA_NodeHelper::CountSiblings() return a vector.
Instead of returning the count of the siblings as an int32_t, and saving
the list of siblings in an in-out parameter. Rename it to GetSiblings().
Adjust the internal TraverseSiblings() helper function to stop counting
as well.

As a result, make GetIndex() return size_t since it also uses
TraverseSiblings() internally. Change many GetIndex() callers to use
size_t as well. Also clean up the CFXJSE_ResolveProcessor methods that
call GetIndex().

Change-Id: Icba3648eb091100beed5162e6b9336de3c122a80