Avoid bad index ASSERT() in CFGAS_StringFormatter::FormatNum
For some reason, the code was concatenating the first character of a
string followed by all the remaining characters of that string. This
is the same as the original string, except when it is empty, in which
case it trips an ASSERT. Just use the original string, as the
computation is pointless.

Bug: 945097
Change-Id: Id26c75d61f17bd858b874f6611d13c889392b071