Retain bitmap in CPDF_RenderStatus::LoadSMask().
Avoids a warning from clang-tidy.

I believe the warning to be spurious since clang-tidy probably
can't reason about the number of other retain pointers to the
underlying object and assumes they might go to zero. Nonetheless,
we're doing enough work here that the cost of holding an additional
reference shouldn't be prohibitive, and it completely avoids the
reliance on other objects keeping the bitmap alive.

Change-Id: I554a5a7171a4c446c67441b1c67b12a7da61b2bc