Update patch set 5
Patch Set 5:

Well, the intra-tree pointers are UnownedPtr<>.  We can't use RetainPtr<> for all the intra-tree pointers because of cycles lead to leaks and we have a cycle between next/prev siblings. We could make FirstChild/NextSibling RetainPtr<> and the other ones UnownedPtr<> to avoid the cycle, and essentially that is what this does, except that less churn is created when assigning to those.

Patch-set: 5
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>