Remove CPDF_FormField::FindOptionValue().
It is identical to FindOption(). At some point in the past,
FindOptionValue() took a starting index parameter (which defaulted
to 0 and was never actually specified in any call) while FindOption()
always started from index 0. But that has long since been removed,
so they are now equivalent.

Pass by const-ref while we're at it.

Change-Id: I3547b8c285b3ee4cbb32a6b4faea06e12bb7e9de