Rename CXFA_ConainerLayoutItem to CXFA_ViewLayoutItem.
Because the names Content/Container are confusingly similar. The
only sub-class is a CFXA_FFPageView, so the name "view" is more
meaningful.

It is doubly confusing because CFXA_Nodes also have a container
property (as in can contain other nodes) that has no relation to
the layout item containers. That property remains named "container".

Change-Id: Ic123599aa10c8656d878e446314e84e9e9a89545