GetFontSize for Variable Text Annotations.
Part of feature work to make Pdfium more usable by mobile platforms that wish to
render comboboxes and textfields using custom displays. Will return Tf font size
for the annotation if applicable. This is obtained from CPDFSDK_Widget because
it is not guaranteed to be set on CPDF_FormField which only sets font
information if DR entry is present.

R=thestig@chromium.org

Bug: b/130297350
Change-Id: Iba866bb47cad78e2aa16343ae8d055887714e6f6