Remove unused argument from CXFA_FFWidget::IsMatchVisibleStatus()
It has been unused ever since XFA was merged into main. There may
have been some confusion between the |dwStatus| argument and the
|m_dwStatus| member, but looking at the call sites, it does not
appear that it would ever be called with an argument that would
cause true to be returned, and everything would be invisible as
a consequence.

Rename to HasVisibleStatus() to indicate that no match against
an argument is to be performed.

Change-Id: I356be74514a80182f91312f1dd5364e3303179b4