Roll third_party/libjpeg_turbo/ 61a2bbaa9..2de84a43e (5 commits)
https://chromium.googlesource.com/chromium/deps/libjpeg_turbo.git/+log/61a2bbaa9aec..2de84a43e683

Update libjpeg-turbo to 2.0.1. This was originally commit 19a3e3aa,
which got reverted in commit c953fde3 because a few PDFs had minor pixel
differences. This turned out to be intentional due to improved
upscaling. So go forward with the roll to match the version of
libjpeg-turbo used in Chromium. The matching Chromium change is:
https://crrev.com/652744

Unlike commit 19a3e3aa, there are no additional changed needed in GN
files, since libjpeg-turbo made all the necessary GN changes already.

TBR=tsepez@chromium.org
BUG=chromium:922430

Created with:
  roll-dep third_party/libjpeg_turbo

Change-Id: I41f50f1f4b849b91039f1525e071456f4461cad1