Provide correct state to pattern rendering.
According to 8.7.3.1 in the PDF 2.0 spec, a tiling pattern should
install the graphics state that was in effect at the beginning of the
pattern’s parent content stream.

This means that when loading the pattern we have to supply the current
graphics state and we also can’t keep the pattern loaded past the
current usage.

Bug:pdfium:1286
Change-Id: Id9f609a52803fec29cb42be2ef9bcda6ff31ce3c