Update patch set 7
Patch Set 7:

> Patch Set 6:
> 
> (1 comment)
> 
> > With the corpus test failing, this CL gets a bit more complicated because we also need to update the corpus and roll DEPS for the corpus. Let me know if that is something you need help with.
> 
> Seems like this test actually does show a problem with my changes. Gonna go back to the drawing board for a little and try to fix this up.

Alright, I think this should finally do it. We should copy all of the states, but set the color state to default because they refer to the current pattern. It is a little difficult for me to make sure all tests succeed now but I *think* this is looking good now. Sorry for going back and forth so often!

Patch-set: 7
Reviewer: Gerrit User 7602 <7602@fca2d64a-0420-3532-b200-12848c95a47b>