Add RetainPtr<>::BackPointer() method.
Return an UnownedPtr from a RetainPtr. Prefered over calls to Get(),
since calls to Get() are almost always avoidable except in this one
circumstance.

Use it in one place and remove some Get calls along the way. In turn,
keep a RetainPtr in one class where we don't need a backpointer.

Change-Id: I89bb6291dc71a62d6aa989858c6e62aaa599ee17