Replace some CPDF_Array::Add() calls with AddNew<>().
Ditto for InsertAt() and SetAt().

AddNew<>() is safer because it can not introduce a direct
cycle of objects into the array since it creates a new object
that can not have references to any old objects [cycles through
indirect objects are always possible, but do not affect memory
management when using RetainPtr<>].

This reduces the number of Add() calls that might need auditing
should a cycle be discovered in arrays again someday.

Change-Id: Ic423fcd4baefed0c692bd60961b360f6bb506e43