Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > On https://bugs.chromium.org/p/pdfium/issues/detail?id=1288, I provided links to a couple PDFs that regressed. bug_1288.in may not fully cover those cases. Can you see if we can avoid those regressions in this CL?
> > 
> > Yep! I'll try to do that soon. Do you remember which PDF you used to base `bug_1288.in` on? http://www.cs.columbia.edu/~lok/3101/lectures/02-corejava.pdf or https://arxiv.org/pdf/1509.05785.pdf?
> 
> Never mind. I can still reproduce the regression on the 1509.05785.pdf. I'll get back to you when this is fixed. Sorry about this!

This is quite curious. The last problem seems to be that there's an optimization in `CPDF_RenderStatus::DrawTilingPattern` at line 2202 which, instead of rendering the pattern into a bitmap simply immediately renders it. If this optimization is hit, it doesn't work right. If we go the normal codepath, it works. Don't quite know *why* yet.

Patch-set: 3
Reviewer: Gerrit User 7602 <7602@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false