Provide correct state to pattern rendering. (try 2)
According to 8.7.3.1 in the PDF 2.0 spec, a tiling pattern should
install the graphics state that was in effect at the beginning of the
pattern’s parent content stream.

The first attempt in commit f188aca283d6dceb874bbe5e20d8003e6bbe184f got
reverted because of an ASAN fuzzer issue.

Those issues have been fixed by not caching the `CPDF_Form` in the
`CPDF_TilingPattern` and simply returning it in a `std::unique_ptr`.

In addition to that the regressions found in bug 1288 have also
been fixed and relevant test cases have been added.

Bug: pdfium:1286
Change-Id: I013dd59587caa9be32963c51934f5030ef1e72fb