Fix ODR violations in public fuzzers
In [1], I assumed a public_fuzzer didn't depend on private PDFium code, however
this is not true.  Some public_fuzzer's have direct dependencies on PDFium
internals.  So in component builds, they will both statically link in parts of
PDFium, and dynamically link in (the same code from) libpdfium.so.

The fix is to export the public fuzzers from libpdfium.so too.  But there's an
annoyance in doing this: the public fuzzers also depend on
pdfium_fuzzer_helper.cc, which depends on test_support.  So test_support and its
dependencies need to be made non-testonly in component builds.

Another annoyance is that pdfium_fuzzer_helper.cc has a static initializer, so
that must be changed to a regular function exported from PDFium so that the real
static initializer in the fuzzers can run it.

[1] https://pdfium.googlesource.com/pdfium/+/93101066fa60961b58324276165a2de6126e0f5d

BUG=960288
R=thestig

Change-Id: I10d03c46c72b454760e5664c4939ab173557cb3b