Fix ODR violations in public fuzzers
In [1], I assumed a public_fuzzer didn't depend on private PDFium code, however
this is not true.  Some public_fuzzer's have direct dependencies on PDFium
internals.  So in component builds, they will both statically link in parts of
PDFium, and dynamically link in (the same code from) libpdfium.so. The fix is
to export the public fuzzers from libpdfium.so too.

[1] https://pdfium.googlesource.com/pdfium/+/93101066fa60961b58324276165a2de6126e0f5d

BUG=960288
R=thestig

Change-Id: I10d03c46c72b454760e5664c4939ab173557cb3b