Fixed issue a line was not displayed when use_skia_paths is true.
When use_skia_paths is true, CFX_SkiaDeviceDriver::DrawPath() was doing
condition checks in a different order compared to
CFX_AggDeviceDriver::DrawPath(), which is used when use_skia_paths flag
is turned off.
This CL fixed the condition check sequence to avoid discrepancy between
the testing results from cases when use_skia_paths is on and off.
Tested on pdf files generated from file bug_1296.in.

Bug:1296
Change-Id: Ie93f8e1055d10210dfe08f53bc20f48d837360d6