Introduce CFX_GEModule::PlatformIface pure virtual class.
This replaces a void* pointer in the CFX_GEModule, and allows platform
destruction order to be controlled by the member order, rather than an
explicit call in the destructor.

Each of the platforms implements its own platform-specific subclass.

In the future, it seems likely that CFX_FontCache/CFX_FontMgr members
will not get to outlive the PlatformIface, and this makes that change
easier.

Change-Id: If278f79b66fa8a7d6b2f9be089c31b37c308ae26