Fix unreadable Japanese PDFs.
There are non-conforming Japanese language PDFs in wide circulation that
use a Type 2 CIDFont including a CIDToGIDMap where the associated True
Type font program is not embedded. The PDF Reference, Sixth Edition,
specifies that a CIDToGIDMap "may appear only in a Type 2 CIDFont whose
associated True Type font program is embedded in the PDF file"
(Section 5.6, pp 437).

These PDFs render correctly if the illegal CIDToGIDMap is ignored. This
patch requires that a CIDFont be Type 2 and associated with an embedded
True Type font before processing any CIDToGIDMap. Otherwise it is
ignored.

BUG=pdfium:1308

Change-Id: I18713f4c0bafcd4a1b601e972cf855c1fea0d0ef