Prevent ciccular includes between fxcodec and fpdfapi/parser.
Hoist streams ever upwards in JBIG2 to remove the dependency. There
is one (set) of retain ptrs removed, but the callers already are
retaining the streams in their own class.

In turn, testing becomes simpler because we don't have to create
phony proxy objects that didn't belong down at this layer in the
first place.

- Rename one member while were at it to more accurately reflect type.
- Remove one test that is no longer relevant since spans don't make
  an empty/null distinction.

Change-Id: I41353a19205e5a32e8428b1d30852d0839996231