Make CPDF_FontGlobals a per-process singleton.
Currently, it is a member of CPDF_PageModule, and while it seems
like a step backwards to re-introduce a global instance, this change
prevents lower-level modules from reaching up into the page module to
get the resources they require. This gets us one step closer to
removing another circular inclusion.

Change-Id: I604cae7c8b240666fa5791a8c8a46e9542e99112