Fixed issue a line was not displayed when use_skia_paths is true.
Using bug_1296.pdf:
The green line wasn't displayed because in
CFX_SkiaDeviceDriver::DrawPath(), when painting a stroke, the check for
"fill_mode & FX_ZEROAREA_FILL" is true, and the rest of drawing step
was skipped.

The black line was not displayed because when
DrawCosmeticLine() is called, pMatrix in DrawPath():line738 is always
null, which result in m_drawMatrix still kept the old value from
previous drawing of the green line.

BUG=pdfium:1296
Change-Id: Ie93f8e1055d10210dfe08f53bc20f48d837360d6