Renove duplicate FWL_VKEY_* definitions.
Add a "XFA_" prefix to the XFA copy in xfa/fwl/fwl_widgetdef.h. Switch
that copy from a set of #defines to an enum, to match the other copy in
public/fpdf_fwlevent.h. Then add assertions that the two copies match,
and update usage in various files. Do IWYU as well.

Change-Id: I4e3c875e4af3380400ba99c1fa9e114e81cd6334