Stop nesting ObservedPtr inside Observable.
Now, only the ObservedPtr<> is templated, and a templated
Observable isn't required. Templated Observable<> lead to
templated ObervedPtr<>::Observables, which made uses of
observed ptr to subclasses awkward, since the Observables
weren't subclasses of each other even if the dependent types
were.

The majority of the change is to fix syntax now that
we no longer have a typename in the classes derived from
Observable.

Change-Id: I59f70915f0a87fa2269d051acff5a7206a3e79e8