Add a pixel test for the PostScript round operator.
PostScript has a special definition for its round operator, which PDFium
does not correctly implement. Add a pixel test with a function-based
shading that uses a PostScript calculator function to exercise the round
operator.

The PDF should render a small yellow rectangle in the lower left corner.
Currently, PDFium incorrectly renders a green rectangle, so suppress the
test result. https://pdfium-review.googlesource.com/56236 is the
follow-up CL that will fix the rendering.

Bug: pdfium:1321
Change-Id: Ib1c262269dbb788073bbd6375411b874c62ae1fd