Make PSOP_ROUND round according to PLRM definiton
Add a function `RoundHalfUp()` to cpdf_psengine.cpp to handle rounding
accoring to the PLRM definition of half up rounding. Half up rounding
of a number `x` is defined as `floor(x + 0.5)`. Support rounding for
all floats, and not only floats within the range of INTs, as before.
This bug is related to pdfium:1314.

Bug: pdfium:1321
Change-Id: I1c071e8ac4bb718a08722159faee6326f82d17ed