[SkiaPath] Fixed an edge case DrawPath() excluded.
For a pdf files that contains the same path more than once and when
fill type was SkPath::kEvenOdd_FillType, the surrounding area was not
filled if the path was drawn even number of times.
To fix the issue, we need to call Flush() to restore SkCanvas every
time a new path is being processed, in regardless whether the new path
is the same as the previous one.

Bug: pdfium:843
Change-Id: Iae9abcddd21180bd49742a02558a95f4506f3d4c