Weakly cache CFX_FontMgr::FontDescs.
The intent is keep the FontDesc alive only so long as there at least
one CFX_Face that is still using it. This is more similar to the
original manually counted implementation from FX than the current
implementation where we keep the FontDesc around for the liftime
of the FontMgr.

Change-Id: If8b441b9b1c2aa81892cb4cb62762a44a322dd0c