Move png encoding logic to helper in pdfium_test_write_helper.cc
The function WriteImages() had a large switch statement for determining
which png encoding to use depending on the format of a bitmap. This
logic can be a separate function that can also be used by WritePng().

Adds success message, "Successfully wrote png <filename>." to
WritePng().

Changes failure message "Failed to open <filename> for output."
to "Failed to open <filename> for saving png."

Change-Id: Ie6148ca560d09b1549feee016eb49055d49abc04