[SkiaPath] Fix issue with missing line.
In bug_1296.pdf, the green line wasn't drawn because in
CFX_SkiaDeviceDriver::DrawPath(), when painting a stroke, the check for
"fill_mode & FX_ZEROAREA_FILL" is always true, and the function exited
and skipped drawing process. Also |CFX_SkiaDeviceDriver| was missing
CFX_SkiaDeviceDriver::GetDriverType(), which resulted in the
discrepancy of the second input for CFX_SkiaDeviceDriver::DrawPath()
when drawing a stroke.
To fix it, need to make CFX_SkiaDeviceDriver::DrawPath() has the same
order of condition checks as CFX_AggDeviceDriver::DrawPath() and the
same input values.


Bug:pdfium:1296
Change-Id: Ie4c318bcf547603edde47a138b48d7a9bd021ff7