Change PNG functions to return std::vector.
Currently, PNG functions pass the std::vector containing the data an an
out parameters, with a separate bool return value to indicate success or
failure. Instead of doing this, just return the std::vector and check
whether it contains data or not.

For affected functions that return std::vector, use the same variable in
all return paths to make it obvious to the compiler that it can apply
RVO and not create a copy.

Change-Id: Ia1f84a6be29fe2d86df800607a258747999b311f