Update patch set 4
Patch Set 4:

It's not obvious if this does the right thing. The best way to determine that is to write a test. FPDFFormFillTextFormEmbedderTest.UndoRedo is a good example to follow. For this case, the test will need to call FORM_ReplaceSelection().

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>