Allow print() only in response to a user gesture
This prevents the print dialog opening automatically when a PDF is
embedded in a web-page.

As part of this change, move the existing tests from document_methods.in
to mouse_events.in, and extend the mouse_events.in tests to check that
print can only be called in response to a user gesture.

BUG=chromium:968914

Change-Id: Ib55808cf500288460f79c940af0c4d2738bb93e5