API to return char indices for weblinks
PDF text is parsed to find links using regex matching. These are not
link which are present as annotation in the PDF but are still
considered links (categorized as Auto-Detected links).
While parsing the links Pdfium identifies the start and end char indices
of the detected link.

This PR exposed a public method to get start and end indices of such
auto detected links. This is required for Accessibility to generate the
relative mapping of links and text.

Modified an existing unittest to cover for both cases: when we ask
indices for a valid link and when we ask for an invalid link.

Bug Link: https://bugs.chromium.org/p/pdfium/issues/detail?id=1334
Change-Id: I6d98b35eb9c6e7fcf4d800776b9129588fce0e09