[SkiaPath] Fixed an edge case DrawPath() excluded.
For a pdf file that contains the same path more than once and when fill
type was an even-odd type, the path's surrounding area is not
filled if the path was drawn for an even number of times.
To fix the issue, we need to make sure DrawChanged() return true when
fill type is an even-odd type, since the number of the same
path showing up will eventually affect the drawing result.

Bug: pdfium:843
Change-Id: Iae9abcddd21180bd49742a02558a95f4506f3d4c