Handle compressed data streams in CPDF_PageContentGenerator.
In CPDF_PageContentGenerator, When it modifies streams,
GenerateModifiedStreams() returns uncompressed data.
UpdateContentStreams() then puts the data into streams. When doing this,
it needs to remember to remove the filters, if any, from the streams.
Otherwise the stream's filters and data may have a mismatch.

Add an embedder test for this case, with yet another variation of
hello_world.pdf.

Bug: pdfium:1326
Change-Id: I59a39cbea1e4d71ed8917c90734e73ca1cdaa411