Bugfix for ToUnicodeMap loading
For the mapping of `bfrange`, the current implementation has a
misunderstanding of the spec. According to the spec (v1.7, 5.9.2,
ToUnicode CMaps, page 474):

 "...ensure that the value of the last byte in the string is
less than or equal to `255 − ( srcCode2 − srcCode1 )`."

The restriction is of the last byte of the Unicode, not the high
char code(srcCode2).

But as a reader here, we should tolerant the mappings that don't
comply with the spec well, so the compliance check is omitted.

Bug: pdfium:1339
Change-Id: I17edde2db86aa406daae5b59bea520cdc71edd4b