Make CPDF_Color saner.
Stop casting a vector of floats to a PatternValue object type. Stop
performing cleanup for said PatternValue outside of the class itself.
Use std::vector in place of mallocs.

In turn, this requires making CPDF_Pattern and its subclasses
retainable, so that the cleanups happen automatically, which is
more straightforward than trying to move the old cleanup logic
into the PatternValue dtor.

In turn, make the doc page data weakly cache the CPDF_Patterns by using
the Observable pattern.

Change-Id: Idd1095cee3bd497392ea1f120bd452f3f34ce104