Update patch set 7
Patch Set 7:

IMHO, if as a PDF generator, we'd better strictly comply with the spec. But as a reader, we'd better tolerant this kind of errors.  My PDF is generated by MS Office, so there must be many docs of this kind in the world and most of the other PDF readers have the tolerance of the errors.

So I removed the compliance check, both of:

- `highcode < lowcode || highcode - lowcode > 255`
- `(last_char & 0xff) > 0xff - (highcode - lowcode)`

What do you think?

Patch-set: 7
Reviewer: Gerrit User 7520 <7520@fca2d64a-0420-3532-b200-12848c95a47b>