Make CLZWDecoder decode in a single pass.
Currently, CLZWDecoder does decoding with 2 passes. The first decode
pass calculates how much space the decoded data takes. Then the caller
has to allocate a buffer of that size. The second decode pass fills in
the buffer with the decoded data.

This CL changes decoding to do all the work in a single pass. The decode
function allocates a small buffer and fills it while decoding. When the
buffer is full, the decode function increases the buffer size
geometrically. This improves overall rendering speed for LZW encoded
PDFs by as much as 4%.

Change-Id: Ieec2e1eda7347a46af434dfba0ac728361d3b5c2