Fix issue with missing thin diagonal lines.
CFX_PathData::GetZeroAreaPath() needs to detect the straight
overlapping paths which forms zero area to fill, and then add the
points which can represent the overlapped path to |NewPath| for
stroking purpose.

The diagonal lines in bug_1338 pixel test are missing because
CFX_PathData::GetZeroAreaPath() only adds points to |NewPath| when the
overlapping paths form vertical or horizontal lines. To fix the issue,
add a check for diagonal lines formed by overlapping paths.

Bug: pdfium:1338
Change-Id: Idd05d7dca437bfedc5a0e75de688c7cc7138d539