Add ASSERTs to prevent re-entrancy inside CPWL_EditImpl_Undo.
In CPWL_EditImpl_Undo, |m_bWorking| is suppose to protect Redo() and
Undo() so nothing gets added in the middle of a redo/undo action. This
only works if Redo() and Undo() are not re-entrant. In the case of
re-entrancy, what can happen is:

1. First Undo() call starts, sets |m_bWorking| to true.
2. Second Undo() call starts, sets |m_bWorking| to true again.
3. Second Undo() call finishes, sets |m_bWorking| to false.
4. First Undo() call finishes, sets |m_bWorking| to false again.

Between events 3 and 4, there is a gap where the first Undo() call has
not finished, and |m_bWorking| is suppose to be true, but has already
been set to false.

Change-Id: I402714f480cb061a5fef83b0c81d10e0a31df83f