Add new experimental API funciton to flatten PDFs without form controls
Add new API function FPDFPage_Flatten_No_Controls that is similar to
the original FPDFPage_Flatten, but takes an additional argument for the
PDF form handle. The form handle allows the flatten function to access
the form controls and cleanly recreate the in-memory widget appearance
streams without having to duplicate code. The original flatten function
calls the new flatten function but passes a nullptr for the form
handle.

Consequently, the CPDF_PageExporter::ExportPage function in fpdf_ppo.cpp
was altered to export the PDF's AcroForm object in the document root.
This is necessary for accessing the PDF's page views and form handles.

Related Chromium CL can be found at [1].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1702627

Bug: chromium:954307

Change-Id: Ia4ab238692bc55bf67f43559c05e1eb2f310b9ff