Let CPDF_Type3Cache directly manage CPDF_CountedFont refcounts.
Currently, this is being perfomred by CPDF_RenderStatus outside
the class. This works because CPDF_RenderStatus::GetCachedType3()
is the only place a CPDF_Type3Cache might get created (and as
a side-effect introduce another unowned pointer into a font managed
elsewhere by a counted font).

Ideally, CPDF_Font would be a Retainable, but there are circular
references to be resolved first. In the mean time, this makes things
a little cleaner.

Change-Id: Id64599f0f4ab6f68f016a0fdaf0e363c45b43808