[SkiaPaths] Avoid overlapping when drawing text.
SkiaPaths always draw text according to each glyph's default width for
a certain font. Sometimes the distances between characters defined in
the pdf file are not as large as the characters' default width, which
causes the characters to overlap. Add process to scale each individual
character according to the ratio of its actual width defined in pdf
file and its default font width to avoid overlapping.

[SkiaPaths] Avoid overlapping when drawing text.

SkiaPaths always draw text according to each glyph's default width for
a certain font. Sometimes the distances between characters defined in
the pdf file are not as large as the characters' default width, which
causes the characters to overlap. Add process to scale each individual
character according to the ratio of its actual width defined in pdf
file and its default font width to avoid overlapping.

Bug:pdfium:1351
Change-Id: I25012829147ef16d06bdbdb57991c80869122f00